backend/requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.20.0
pydantic>=1.10.0
python-multipart>=0.0.6
PyPDF2>=3.0.0
nltk>=3.8.0
google-generativeai>=0.2.0
requests>=2.28.0
python-dotenv>=0.21.0
